cleanup: remove commented code, rolling hash interface changes
